requirements.txt
art
ascii_magic==2.3.0
tabulate
mistralai
python-dotenv
pydantic
moviepy
pytube
pytubefix
peewee
flask
pyTelegramBotAPI
python-telegram-bot
flask_restx

# .\.venv\Scripts\activate - активирует виртуальное окружение
# pip install -r requirements.txt - устанавливает все зависимости из файла requirements.txt
# python main.py - запускает программу
# python -m pip install --upgrade pip - обновляет pip
# pip install --upgrade pip - обновляет pip
# pip install moviepy
# pip install moviepy pytubefix
# git push origin main - отправляет изменения в репозиторий
# pip install peewee  
# pip install flask
# git fetch - обновляет репозиторий
# python -m venv .venv - создает виртуальное окружение
# pip install pyTelegramBotAPI - устанавливает библиотеку pyTelegramBotAPI
# pip install python-telegram-bot - телеграмм бот
# pip install flask_restx - 


